backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic
openai
python-whois
slowapi
stripe
